requirements.txt
flask
pandas
python-dotenv
google-generativeai